requirements.txt
ansible==2.5.4
asn1crypto==0.24.0
bcrypt==3.1.4
cffi==1.11.5
cryptography==2.2.2
Django==2.0.5
djangorestframework==3.8.2
idna==2.6
Jinja2==2.10
MarkupSafe==1.0
paramiko==2.4.1
psycopg2==2.7.4
psycopg2-binary==2.7.4
pyasn1==0.4.3
pycparser==2.18
PyNaCl==1.2.1
pytz==2018.4
PyYAML==3.12
six==1.11.0
